+++ NEWS.md
@@ -1,5 +1,7 @@
 # rsi (development version)
 
+* `get_stac_data()` now provides a more informative error when 0 items are found for a given query.
+
 # rsi 0.1.0
 
 * Initial CRAN submission.